fix(auth): implement jwks cache and refresh mechanism
